훅에서 너무 많은 로직을 처리하고 있어서 분리 (WIP)
